+++ src/Router.js
@@ -4,7 +4,6 @@ import Login from './pages/Login/Login';
 import Main from './pages/Main/Main';
 import ProductList from './pages/ProductList/ProductList';
 import ProductDetail from './pages/ProductDetail/ProductDetail';
-import CategoryList from './pages/ProductList/CategoryList/CategoryList';
 
 function Router() {
   return (
@@ -12,9 +11,7 @@ function Router() {
       <Routes>
         <Route path="/login" element={<Login />} />
         <Route path="/" element={<Main />} />
-        <Route path="/product-list" element={<ProductList />}>
-          <Route path=":category" element={<CategoryList />} />
-        </Route>
+        <Route path="/product-list" element={<ProductList />} />
         <Route path="/product-detail" element={<ProductDetail />} />
       </Routes>
     </BrowserRouter>
