+++ src/pages/ProductList/FIlter/Filter.scss
@@ -1,0 +1,56 @@
+@import '../../../styles/variables.scss';
+
+.filter {
+  @include center($justify-content: space-between);
+  padding: 23px 40px 23px 20px;
+  background-color: #ebeade;
+
+  .filterSubNavContainer {
+    @include center($justify-content: space-between);
+    padding-left: 10px;
+
+    .filterSubNavList {
+      padding: 0 10px;
+      font-size: $font-small;
+      color: $font-light-gray;
+
+      &:first-child {
+        margin-right: 5px;
+        padding-right: 10px;
+        border-right: 1px solid $font-light-gray;
+      }
+
+      .filterSubNavBtn {
+        font-size: $font-small;
+        color: $font-light-gray;
+      }
+    }
+  }
+  .filterBtn {
+    padding: 21px 10px;
+    border: 1px solid rgba(0, 0, 0, 0.2);
+    transition: 0.2s ease-in;
+
+    &:hover {
+      background-color: $color-dark-gray;
+
+      .filterBtnText,
+      .filterBtnIcon {
+        color: $color-beige;
+      }
+    }
+
+    .filterBtnText,
+    .filterBtnIcon {
+      color: $font-light-gray;
+      font-size: $font-small;
+      margin: 0 14px;
+    }
+  }
+  .filterXBtn {
+    padding: 20px 10px;
+    .fa-xmark {
+      font-size: 20px;
+    }
+  }
+}
