+++ src/pages/ProductList/ProductList.scss
@@ -1,84 +1,8 @@
 @import '../../styles/variables.scss';
 
-.ProductList {
-  background-color: $color-beige;
-  button {
-    padding: 0;
-    background-color: transparent;
-    outline: none;
-    border: none;
-    cursor: pointer;
-  }
-
-  header {
-    margin: 100px 0 20px;
-
-    .logo {
-      width: 130px;
-      margin-left: 25px;
-    }
-
-    .mainCategory {
-      margin: 70px 0 20px 0;
-      padding: 0 20px 0 40px;
-      font-size: $font-medium;
-      color: $color-dark-gray;
-    }
-  }
-
-  .filterWrapper {
-    @include center($justify-content: space-between);
-    padding: 23px 40px 23px 20px;
-    background-color: #ebeade;
-
-    .filterSubNavContainer {
-      @include center($justify-content: space-between);
-      padding-left: 10px;
-
-      .filterSubNavList {
-        padding: 0 10px;
-        font-size: $font-small;
-        color: $font-light-gray;
-        text-decoration: none;
-
-        &:first-child {
-          margin-right: 5px;
-          padding-right: 10px;
-          border-right: 1px solid $font-light-gray;
-        }
-
-        .filterSubNavLink {
-          font-size: $font-small;
-          color: $font-light-gray;
-        }
-      }
-    }
-    .filterBtn {
-      padding: 21px 10px;
-      border: 1px solid rgba(0, 0, 0, 0.2);
-      transition: 0.2s ease-in;
-
-      &:hover {
-        background-color: $color-dark-gray;
-
-        .filterBtnText,
-        .filterBtnIcon {
-          color: $color-beige;
-        }
-      }
-
-      .filterBtnText,
-      .filterBtnIcon {
-        color: $font-light-gray;
-        font-size: $font-small;
-        margin: 0 14px;
-      }
-    }
-    .filterXBtn {
-      padding: 20px 10px;
-      .fa-xmark {
-        font-size: 20px;
-      }
-    }
-  }
+.mainCategory {
+  margin: 70px 0 20px 0;
+  padding: 0 20px 0 40px;
+  font-size: $font-medium;
+  color: $color-dark-gray;
 }
