+++ src/pages/ProductList/ProductListLayout/ProductListLayout.scss
@@ -1,0 +1,21 @@
+@import '../../../styles/variables.scss';
+
+.ProductList {
+  background-color: $color-beige;
+  button {
+    padding: 0;
+    background-color: transparent;
+    outline: none;
+    border: none;
+    cursor: pointer;
+  }
+
+  header {
+    margin: 100px 0 20px;
+
+    .logo {
+      width: 130px;
+      margin-left: 25px;
+    }
+  }
+}
